--- src/common/ui/ImageViewer.css
@@ -1,97 +1,0 @@
-.ImageViewer {
-  position: relative;
-  width: 100%;
-  min-height: min(72vh, 32rem);
-  margin-top: 0.5rem;
-  background: #f6f7f8;
-  border: 1px solid #d8dde3;
-  border-radius: 4px;
-  overflow: hidden;
-}
-
-.ImageViewer--empty {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: #5c6670;
-}
-
-.ImageViewer__toolbar {
-  position: absolute;
-  top: 0.75rem;
-  left: 0.75rem;
-  z-index: 2;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.ImageViewer__control {
-  min-width: 2.5rem;
-  min-height: 2.5rem;
-  border-color: #c3cad2;
-  color: #1f2933;
-  background-color: rgb(255 255 255 / 94%);
-}
-
-.ImageViewer__control:focus-visible {
-  outline: 3px solid #7aa7d9;
-  outline-offset: 2px;
-}
-
-.ImageViewer__scale {
-  min-width: 3.5rem;
-  padding: 0.45rem 0.65rem;
-  border: 1px solid #c3cad2;
-  border-radius: 4px;
-  color: #1f2933;
-  background-color: rgb(255 255 255 / 94%);
-  font-variant-numeric: tabular-nums;
-  text-align: center;
-}
-
-.ImageViewer .react-transform-wrapper,
-.ImageViewer .react-transform-component,
-.ImageViewer__viewport {
-  width: 100%;
-  max-width: 100%;
-}
-
-.ImageViewer .react-transform-component,
-.ImageViewer__viewport {
-  min-height: min(72vh, 32rem);
-  max-height: 72vh;
-  overflow: hidden;
-  touch-action: none;
-}
-
-.ImageViewer__content {
-  width: 100%;
-  min-height: min(72vh, 32rem);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.ImageViewer__image {
-  display: block;
-  width: auto;
-  height: auto;
-  max-width: 100%;
-  max-height: 72vh;
-  object-fit: contain;
-  user-select: none;
-}
-
-@media (max-width: 575.98px) {
-  .ImageViewer__toolbar {
-    right: 0.5rem;
-    left: 0.5rem;
-    flex-wrap: wrap;
-  }
-
-  .ImageViewer__control {
-    min-width: 2.75rem;
-    min-height: 2.75rem;
-  }
-}
