+++ src/fragmentarium/infrastructure/mediaMapperValidation.ts
@@ -1,0 +1,30 @@
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null
+}
+
+export function normalizeNonEmptyString(value: unknown): string | undefined {
+  if (typeof value !== 'string') {
+    return undefined
+  }
+
+  const normalizedValue = value.trim()
+  return normalizedValue === '' ? undefined : normalizedValue
+}
+
+export function normalizeNonNegativeInteger(
+  value: unknown,
+): number | undefined {
+  return Number.isInteger(value) && (value as number) >= 0
+    ? (value as number)
+    : undefined
+}
+
+export function normalizePositiveInteger(value: unknown): number | undefined {
+  return Number.isInteger(value) && (value as number) > 0
+    ? (value as number)
+    : undefined
+}
+
+export function normalizeOptionalText(value: unknown): string | undefined {
+  return normalizeNonEmptyString(value)
+}
